reorganized parser/quads source files into subdirs
