+++ include/parser/scope.h
@@ -10,7 +10,7 @@
 
 #ifndef SCOPEH
 #define SCOPEH
-#include <symtab.h>
+#include <parser/symtab.h>
 
 enum name_space { NS_TAG, NS_LABEL, NS_IDENT };
 enum scope_type { ST_FILE, ST_FUNC, ST_BLOCK, ST_PROTO, ST_STRUCTUNION };
