+++ src/parser/symtab.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <symtab.h>
+#include <parser/symtab.h>
 
 // good hash primes: https://planetmath.org/goodhashtableprimes
 static int ghp[] = {53, 97, 193, 389, 769, 1533, 3079, 6151, 12289, 24593};
